backend/requirements.txt
fastapi
uvicorn[standard]
httpx
neo4j
python-dotenv
networkx
